+++ CHANGELOG.md
@@ -31,6 +31,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - In `POST /api/v1/wallet/transaction`, moved `wallet` parameters to the top level of the object
 - Incoming wire message size limit increased to 1024kB
 - Clients restrict the maximum number of blocks they will send in a `GiveBlocksMessage` to 20
+- `POST /api/v2/wallet/seed/verify` returns an error if the seed's checksum is invalid
+- Increase the detail of error messages for invalid seeds sent to `POST /api/v2/wallet/seed/verify`
+- Move package `github.com/skycoin/skycoin/src/cipher/go-bip39` to `github.com/skycoin/skycoin/src/cipher/bip39`
 
 ### Removed